refactor!: move away from loading package files
This library was a little overcomplicated and it doesn't have a single
responsibility. I've stripped it back a lot: the package no longer loads
the package files from the file system, instead you're required to pass
in a pre-loaded package file.

This library now plays really nicely with @rowanmanning/package-json.
It's also doing _just_ enough work - not leaving the work of navigating
through package properties to the user.